requirements.txt
flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
Flask-Mail==0.9.1 